% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{call_mixed}
\alias{call_mixed}
\title{Run C and Fortran code
#'}
\usage{
call_mixed(x)
}
\arguments{
\item{x}{A single integer to be passed into C/Fortran.}
}
\description{
Run C and Fortran code
#'
}
\examples{
\donttest{
# Example C/Fortran run

call_mixed(
  x = 12
)
}
}
